Reestructuración paquete domain, con routePro
clases que se conservan en routePro

DistanceTime
DistanceTimeData
Location

+++ base project/src/main/java/cl/citymovil/route_pro/message_listener/domain/DistanceTime.java
@@ -25,8 +25,8 @@ public class DistanceTime {
 	
 	private long destination;
 	
-	private double distance;
-	private double duration;
+	private long distance;
+	private long duration;
 	
 	public DistanceTime(){}
 
@@ -63,7 +63,7 @@ public class DistanceTime {
 		this.destination = destination;
 	}
 
-	public double getDistance() {
+	public long getDistance() {
 		return distance;
 	}
 
@@ -71,7 +71,7 @@ public class DistanceTime {
 		this.distance = distance;
 	}
 
-	public double getDuration() {
+	public long getDuration() {
 		return duration;
 	}
